refactor(database): extract user model + testing
