+++ lib/services/database/firestore_service.dart
@@ -1,0 +1,7 @@
+import "package:cloud_firestore/cloud_firestore.dart";
+import "package:hooks_riverpod/hooks_riverpod.dart";
+
+/// Static singleton [FirebaseFirestore] instance provider
+final firestoreProvider = Provider<FirebaseFirestore>(
+  (ref) => FirebaseFirestore.instance,
+);
