+++ test/models/database/mock_post_data.dart
@@ -1,7 +1,7 @@
 import "package:cloud_firestore/cloud_firestore.dart";
 import "package:geoflutterfire2/geoflutterfire2.dart";
 import "package:proxima/models/database/post_repository.dart";
-import "package:proxima/models/database/user_repository.dart";
+import "package:proxima/models/database/user/user_id_firestore.dart";
 
 class MockFirestorePost {
   PostFirestore createPostAt(
